fix(ui): horloge de la barre supérieure en gras, mois abrégé
- digitalClock : mois en format court (« sept. », « janv. ») pour raccourcir la pilule.
- _Layout : date et heure en font-semibold (text-slate-700) pour la lisibilité.

+++ src/SamaEcole.Web/wwwroot/js/ui-components.js
@@ -495,10 +495,11 @@ document.addEventListener('alpine:init', () => {
         },
         updateClock() {
             const now = new Date();
+            // Mois abrégé (« sept. », « janv. »…) : la pilule reste compacte sans rogner le jour de la semaine.
             const dateStr = now.toLocaleDateString('fr-FR', {
                 weekday: 'long',
                 year: 'numeric',
-                month: 'long',
+                month: 'short',
                 day: 'numeric'
             });
             const timeStr = now.toLocaleTimeString('fr-FR', {
